+++ src/comms/display.py
@@ -31,22 +31,22 @@ def show(message):
 def status_summary():
     if comms_node.getBool("display_on"):
         print("[imu  ]:gyro = %.3f %.3f %.3f [deg/s]" % \
-              (imu_node.getFloat("p_rad_sec") * r2d,
-               imu_node.getFloat("q_rad_sec") * r2d,
-               imu_node.getFloat("r_rad_sec") * r2d),
+              (imu_node.getDouble("p_rad_sec") * r2d,
+               imu_node.getDouble("q_rad_sec") * r2d,
+               imu_node.getDouble("r_rad_sec") * r2d),
               "accel = %.3f %.3f %.3f [m/s^2]" % \
-              (imu_node.getFloat("ax_mps_sec"),
-               imu_node.getFloat("ay_mps_sec"),
-               imu_node.getFloat("az_mps_sec")))
+              (imu_node.getDouble("ax_mps_sec"),
+               imu_node.getDouble("ay_mps_sec"),
+               imu_node.getDouble("az_mps_sec")))
         print("[mag  ]:%.3f %.3f %.3f" % \
-              (imu_node.getFloat("hx"),
-               imu_node.getFloat("hy"),
-               imu_node.getFloat("hz")))
+              (imu_node.getDouble("hx"),
+               imu_node.getDouble("hy"),
+               imu_node.getDouble("hz")))
         print("[air  ]:Palt = %5.2f[m] Pspd = %4.1f[kt]" % \
-              (pos_pressure_node.getFloat("altitude_m"),
-               velocity_node.getFloat("airspeed_kt")))
+              (pos_pressure_node.getDouble("altitude_m"),
+               velocity_node.getDouble("airspeed_kt")))
 
-        if gps_node.getFloat("data_age") < 10.0:
+        if gps_node.getDouble("data_age") < 10.0:
             print("[gps  ]:date = %04d/%02d/%02d %02d:%02d:%02d" % \
                   (gps_node.getInt("year"),
                    gps_node.getInt("month"),
@@ -57,32 +57,32 @@ def status_summary():
             print("[gps  ]: pos = %.6f %.6f %.1fm sats = %ld, age = %.2f" % \
                   (gps_node.getDouble("longitude_deg"),
                    gps_node.getDouble("latitude_deg"),
-                   gps_node.getFloat("altitude_m"),
+                   gps_node.getDouble("altitude_m"),
                    gps_node.getInt("satellites"),
-                   gps_node.getFloat("data_age")))
+                   gps_node.getDouble("data_age")))
         else:
-            print("[gps  ]: age =", gps_node.getFloat("data_age"))
+            print("[gps  ]: age =", gps_node.getDouble("data_age"))
 
         filter_status = status_node.getString("navigation")
         print("[filt ]:[%s] " % filter_status, end="")
         print("pos = %.6f %.6f %.1f " % \
               (filter_node.getDouble("longitude_deg"),
                filter_node.getDouble("latitude_deg"),
-               filter_node.getFloat("altitude_m")), end = "")
+               filter_node.getDouble("altitude_m")), end = "")
         print("RPY = %4.1f %4.1f %5.1f (deg)" % \
-              (orient_node.getFloat("roll_deg"),
-               orient_node.getFloat("pitch_deg"),
-               orient_node.getFloat("heading_deg")))
+              (orient_node.getDouble("roll_deg"),
+               orient_node.getDouble("pitch_deg"),
+               orient_node.getDouble("heading_deg")))
 
         print("[act  ]:%.2f %.2f %.2f %.2f %.2f" % \
-              (act_node.getFloat("aileron"),
-               act_node.getFloat("elevator"),
-               act_node.getFloat("throttle"),
-               act_node.getFloat("rudder"),
-               act_node.getFloat("flaps")))
+              (act_node.getDouble("aileron"),
+               act_node.getDouble("elevator"),
+               act_node.getDouble("throttle"),
+               act_node.getDouble("rudder"),
+               act_node.getDouble("flaps")))
         print("[hlth ]:cmdseq = %ld  tgtwp = %ld  loadavg = %.2f  vcc = %.2f" % \
               (remote_link_node.getInt("sequence_num"),
                route_node.getInt("target_waypoint_idx"),
-               status_node.getFloat("system_load_avg"),
-               power_node.getFloat("avionics_vcc")))
+               status_node.getDouble("system_load_avg"),
+               power_node.getDouble("avionics_vcc")))
         print()
